:refactoring: pretty objects 2

+++ internal/service/logic/logic.go
@@ -2,7 +2,6 @@ package logic
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"github.com/egorgasay/gost"
 	"go.uber.org/zap"
@@ -10,7 +9,6 @@ import (
 	"itisadb/internal/constants"
 	"itisadb/internal/domains"
 	"itisadb/internal/models"
-	"sync"
 )
 
 type Logic struct {
@@ -31,21 +29,29 @@ func NewLogic(storage domains.Storage, cfg config.Config, tlogger domains.Transa
 	}
 }
 
-func (l *Logic) Find(ctx context.Context, key string, out chan<- string, once *sync.Once, opts models.GetOptions) {
-	//TODO implement me
-	panic("implement me")
-}
-
-func (l *Logic) GetOne(ctx context.Context, key string, opt models.GetOptions) (res gost.Result[string]) {
+func (l *Logic) GetOne(_ context.Context, userID int, key string, _ models.GetOptions) (res gost.Result[models.Value]) {
 	v, err := l.storage.Get(key)
 	if err != nil {
 		return res.ErrNew(0, 0, err.Error())
 	}
 
-	return res.Ok(v.Value)
+	if !l.hasPermission(userID, v.Level) {
+		return res.Err(constants.ErrForbidden)
+	}
+
+	return res.Ok(v)
 }
 
-func (l *Logic) DelOne(ctx context.Context, key string, opt models.DeleteOptions) (res gost.Result[gost.Nothing]) {
+func (l *Logic) DelOne(_ context.Context, userID int, key string, _ models.DeleteOptions) (res gost.Result[gost.Nothing]) {
+	v, err := l.storage.Get(key)
+	if err != nil {
+		return res.ErrNew(0, 0, err.Error())
+	}
+
+	if !l.hasPermission(userID, v.Level) {
+		return res.Err(constants.ErrForbidden)
+	}
+
 	if err := l.storage.Delete(key); err != nil {
 		l.logger.Warn("failed to delete", zap.Error(err))
 		return res.ErrNew(0, 0, err.Error())
@@ -58,10 +64,20 @@ func (l *Logic) DelOne(ctx context.Context, key string, opt models.DeleteOptions
 	return res.Ok(gost.Nothing{})
 }
 
-func (l *Logic) SetOne(ctx context.Context, key string, val string, opt models.SetOptions) (res gost.Result[int32]) {
+func (l *Logic) SetOne(_ context.Context, userID int, key string, val string, opt models.SetOptions) (res gost.Result[int32]) {
+	if !l.hasPermission(userID, opt.Level) {
+		return res.Err(constants.ErrForbidden)
+	}
+
 	v, err := l.storage.Get(key)
-	if err == nil && (opt.Unique || v.ReadOnly) {
-		return res.Err(constants.ErrAlreadyExists)
+	if err == nil {
+		if !l.hasPermission(userID, v.Level) {
+			return res.Err(constants.ErrForbidden)
+		}
+
+		if opt.Unique || v.ReadOnly {
+			return res.Err(constants.ErrAlreadyExists)
+		}
 	}
 
 	err = l.storage.Set(key, val, opt)
@@ -76,28 +92,18 @@ func (l *Logic) SetOne(ctx context.Context, key string, val string, opt models.S
 	return res.Ok(constants.MainStorageNumber)
 }
 
-func (l *Logic) earlyObjectNotFound(requested *int32, actual int32) bool {
-	if requested != nil {
-		return *requested != actual
+func (l *Logic) HasPermissionToObject(userID int, name string) (res gost.Result[bool]) {
+	info, err := l.storage.GetObjectInfo(name)
+	if err != nil {
+		return res.ErrNew(0, 0, err.Error())
 	}
 
-	return false
+	return res.Ok(l.hasPermission(userID, info.Level))
 }
 
-func (l *Logic) NewObject(ctx context.Context, name string, opts models.ObjectOptions) (res gost.ResultN) {
-	info, err := l.storage.GetObjectInfo(name)
-	errNotFound := errors.Is(err, constants.ErrNotFound)
-
-	if err != nil && !errNotFound {
-		return 0, fmt.Errorf("can't get object info: %w", err)
-	}
-
-	if !errNotFound {
-		if info.Server != _mainStorage {
-			return 0, constants.ErrServerNotFound
-		}
-
-		return _mainStorage, nil
+func (l *Logic) NewObject(_ context.Context, userID int, name string, opts models.ObjectOptions) (res gost.Result[gost.Nothing]) {
+	if !l.hasPermission(userID, opts.Level) {
+		return res.Err(constants.ErrForbidden)
 	}
 
 	if err := l.storage.CreateObject(name, opts); err != nil {
@@ -109,22 +115,52 @@ func (l *Logic) NewObject(ctx context.Context, name string, opts models.ObjectOp
 		Level:  opts.Level,
 	}
 
-	l.storage.AddObjectInfo(name, info)
+	l.storage.AddObjectInfo(name, info) // TODO: maybe you should union Create + AddObjectInfo? and keep all information about object in one place?
 
 	if l.cfg.TransactionLogger.On {
 		l.tlogger.WriteCreateObject(name)
 		l.tlogger.WriteAddObjectInfo(name, info)
 	}
 
-	return res.Ok()
+	return res.Ok(gost.Nothing{})
 }
 
-func (l *Logic) SetToObject(ctx context.Context, object string, key string, value string, opts models.SetToObjectOptions) (res gost.Result[gost.Nothing]) {
-	//TODO implement me
-	panic("implement me")
+func (l *Logic) SetToObject(_ context.Context, userID int, object string, key string, value string, opts models.SetToObjectOptions) (res gost.Result[gost.Nothing]) {
+	info, err := l.storage.GetObjectInfo(object)
+	if err != nil {
+		return res.ErrNew(0, 0, err.Error()) // TODO: ??
+	}
+
+	if !l.hasPermission(userID, info.Level) {
+		return res.Err(constants.ErrForbidden)
+	}
+
+	err = l.storage.SetToObject(object, key, value, opts)
+	if err != nil {
+		return res.ErrNew(0, 0, err.Error())
+	}
+
+	if l.cfg.TransactionLogger.On {
+		l.tlogger.WriteSetToObject(object, key, value)
+	}
+
+	return res.Ok(gost.Nothing{})
 }
 
-func (l *Logic) GetFromObject(ctx context.Context, object string, key string, opts models.GetFromObjectOptions) (res gost.Result[string]) {
-	//TODO implement me
-	panic("implement me")
+func (l *Logic) GetFromObject(_ context.Context, userID int, object, key string, _ models.GetFromObjectOptions) (res gost.Result[string]) {
+	info, err := l.storage.GetObjectInfo(object)
+	if err != nil {
+		return res.ErrNew(0, 0, err.Error())
+	}
+
+	if !l.hasPermission(userID, info.Level) {
+		return res.Err(constants.ErrForbidden)
+	}
+
+	v, err := l.storage.GetFromObject(object, key)
+	if err != nil {
+		return res.ErrNew(0, 0, err.Error())
+	}
+
+	return res.Ok(v)
 }
